DeepHyperion+SaFliTe/DeepHyperion+SaFliTe_mistral/requirements.txt
aiogrpc==1.8
catkin-pkg==1.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
csv-logger==1.3.0
cycler==0.12.1
Cython==3.0.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
docutils==0.20.1
fonttools==4.43.1
grpcio==1.59.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
joblib==1.3.2
keyboard==0.13.5
kiwisolver==1.4.5
llvmlite==0.41.0
lxml==4.9.3
matplotlib==3.7.3
mavsdk==1.4.1
munch==2.5.0
numba==0.58.0
numpy==1.23.2
packaging==23.2
pandas==2.0.3
patsy==0.5.3
Pillow==10.1.0
protobuf==4.24.4
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3.post1
pyulog==0.9.0
PyYAML==6.0
requests==2.31.0
rospkg==1.4.0
ruptures==1.1.7
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
Shapely==1.8.4
similaritymeasures==0.5.0
six==1.16.0
statsmodels==0.13.2
threadpoolctl==3.2.0
tslearn==0.5.2
tzdata==2023.3
urllib3==2.0.6
validators==0.20.0
webdavclient3==3.14.6
zipp==3.17.0
transformers
llama-cpp-python

